refactor(types,shared,elements): Assign `telemetry` property to main Clerk interface (#3329)
* refactor(types,shared): Create separate type interface for `TelemetryCollector`

* fix(shared): Remove `TelemetryEvent` interfaces

* fix(elements): Remove assertion for `telemetry`

* docs(types): Add JSDocs on telemetry types to indicate internal usage

* refactor(elements): Remove `useTelemetry` hook

* Update .changeset/heavy-kings-fix.md

---------
