+++ packages/elements/src/react/sign-in/root.tsx
@@ -10,7 +10,6 @@ import { SignInRouterMachine } from '~/internals/machines/sign-in';
 import { consoleInspector } from '~/internals/utils/inspector';
 import { Router, useClerkRouter, useNextRouter } from '~/react/router';
 import { SignInRouterCtx } from '~/react/sign-in/context';
-import { useTelemetry } from '~/react/utils/telemetry';
 
 import { Form } from '../common/form';
 
@@ -83,9 +82,9 @@ export function SignInRoot({
   fallback = null,
   exampleMode,
 }: SignInRootProps): JSX.Element | null {
-  const telemetry = useTelemetry();
+  const clerk = useClerk();
 
-  telemetry?.record(
+  clerk.telemetry?.record(
     eventComponentMounted('Elements_SignInRoot', {
       path,
       fallback: Boolean(fallback),
