+++ packages/elements/src/react/sign-up/root.tsx
@@ -11,7 +11,6 @@ import { SignUpRouterMachine } from '~/internals/machines/sign-up';
 import { consoleInspector } from '~/internals/utils/inspector';
 import { Router, useClerkRouter, useNextRouter } from '~/react/router';
 import { SignUpRouterCtx } from '~/react/sign-up/context';
-import { useTelemetry } from '~/react/utils/telemetry';
 
 import { Form } from '../common/form';
 
@@ -80,9 +79,9 @@ export function SignUpRoot({
   fallback = null,
   exampleMode,
 }: SignUpRootProps): JSX.Element | null {
-  const telemetry = useTelemetry();
+  const clerk = useClerk();
 
-  telemetry?.record(
+  clerk.telemetry?.record(
     eventComponentMounted('Elements_SignUpRoot', {
       path,
       fallback: Boolean(fallback),
