--- packages/elements/src/react/utils/telemetry.ts
@@ -1,8 +1,0 @@
-import { useClerk } from '@clerk/clerk-react';
-import type { TelemetryCollector } from '@clerk/shared/telemetry';
-
-export const useTelemetry = () => {
-  // @ts-expect-error - The telemetry property is just missing from LoadedClerk :/
-  const { telemetry } = useClerk();
-  return telemetry as TelemetryCollector | undefined;
-};
