+++ packages/shared/src/telemetry/collector.ts
@@ -10,12 +10,17 @@
  *
  * For more information, please see the telemetry documentation page: https://clerk.com/docs/telemetry
  */
-import type { InstanceType } from '@clerk/types';
+import type {
+  InstanceType,
+  TelemetryCollector as TelemetryCollectorInterface,
+  TelemetryEvent,
+  TelemetryEventRaw,
+} from '@clerk/types';
 
 import { parsePublishableKey } from '../keys';
 import { isTruthy } from '../underscore';
 import { TelemetryEventThrottler } from './throttler';
-import type { TelemetryCollectorOptions, TelemetryEvent, TelemetryEventRaw } from './types';
+import type { TelemetryCollectorOptions } from './types';
 
 type TelemetryCollectorConfig = Pick<
   TelemetryCollectorOptions,
@@ -42,7 +47,7 @@ const DEFAULT_CONFIG: Partial<TelemetryCollectorConfig> = {
   endpoint: 'https://clerk-telemetry.com',
 };
 
-export class TelemetryCollector {
+export class TelemetryCollector implements TelemetryCollectorInterface {
   #config: Required<TelemetryCollectorConfig>;
   #eventThrottler: TelemetryEventThrottler;
   #metadata: TelemetryMetadata = {} as TelemetryMetadata;
