+++ packages/types/src/index.ts
@@ -15,6 +15,7 @@ export * from './factors';
 export * from './identificationLink';
 export * from './identifiers';
 export * from './image';
+export * from './instance';
 export * from './json';
 export * from './jwt';
 export * from './key';
@@ -45,6 +46,7 @@ export * from './strategies';
 export * from './theme';
 export * from './token';
 export * from './totp';
+export * from './telemetry';
 export * from './user';
 export * from './userOrganizationInvitation';
 export * from './userSettings';
