+++ packages/types/src/instance.ts
@@ -1,0 +1,1 @@
+export type InstanceType = 'production' | 'development';
