+++ packages/types/src/telemetry.ts
@@ -1,0 +1,48 @@
+import type { InstanceType } from 'instance';
+
+/**
+ * @internal
+ */
+export type TelemetryEvent = {
+  event: string;
+  /**
+   * publishableKey
+   */
+  pk?: string;
+  /**
+   * secretKey
+   */
+  sk?: string;
+  /**
+   * instanceType
+   */
+  it: InstanceType;
+  /**
+   * clerkVersion
+   */
+  cv: string;
+  /**
+   * SDK
+   */
+  sdk?: string;
+  /**
+   * SDK Version
+   */
+  sdkv?: string;
+  payload: Record<string, string | number | boolean>;
+};
+
+/**
+ * @internal
+ */
+export type TelemetryEventRaw<Payload = TelemetryEvent['payload']> = {
+  event: TelemetryEvent['event'];
+  eventSamplingRate?: number;
+  payload: Payload;
+};
+
+export interface TelemetryCollector {
+  isEnabled: boolean;
+  isDebug: boolean;
+  record(event: TelemetryEventRaw): void;
+}
